+++ app/src/main/java/co/develhope/chooseyouowncocktail_g2/DrinkAction.kt
@@ -1,0 +1,8 @@
+package co.develhope.chooseyouowncocktail_g2
+
+import co.develhope.chooseyouowncocktail_g2.model.Beer
+
+sealed class DrinkAction {
+    data class GotoDetail(val beer: Beer) : DrinkAction()
+    object SetPref : DrinkAction()
+}
