+++ app/src/main/java/co/develhope/chooseyouowncocktail_g2/MainActivity.kt
@@ -26,7 +26,6 @@ class MainActivity : AppCompatActivity() {
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        Action.activity = this
 
         val navView: BottomNavigationView = binding.navView
